requirements.txt
torch
torchvision
numpy
matplotlib
seaborn
Pillow
tqdm
diffusers
transformers
accelerate